\name{bean.data}
\alias{bean.data}
\docType{data}
\title{
Bean Data
}
\description{
A simulated bean yield data.frame
}
\usage{data(bean.data)}
\format{
  A data frame with 20 genotypes on the following 10 environments.
  \describe{
    \item{\code{amb1}}{a numeric vector}
    \item{\code{amb2}}{a numeric vector}
    \item{\code{amb3}}{a numeric vector}
    \item{\code{amb4}}{a numeric vector}
    \item{\code{amb5}}{a numeric vector}
    \item{\code{amb6}}{a numeric vector}
    \item{\code{amb7}}{a numeric vector}
    \item{\code{amb8}}{a numeric vector}
    \item{\code{amb9}}{a numeric vector}
    \item{\code{amb10}}{a numeric vector}
  }
}

\examples{
data(bean.data)
## maybe str(bean.data) ; plot(bean.data) ...
}
\keyword{datasets}
